requirements.txt
openai
flask
apscheduler
python-dotenv
